backend/flask-services/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
eventlet==0.39.0
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.10.1
cryptography==44.0.0
boto3==1.36.8
botocore==1.36.8
pymongo==4.11.1
redis==5.2.1
faiss-cpu==1.11.0
numpy==2.3.1
nltk==3.9.1
PyYAML==6.0.2